Attempt to fix code climate warnings
Rubocop warning

Refactor helper
